Move robot to FreeRTOS
Inside the FreeRTOS zone, we can implement cli commands for dealing
with the robot.
